+++ sdk/healthcareapis/azure-resourcemanager-healthcareapis/src/main/java/com/azure/resourcemanager/healthcareapis/implementation/DicomServiceImpl.java
@@ -18,6 +18,7 @@ import com.azure.resourcemanager.healthcareapis.models.ProvisioningState;
 import com.azure.resourcemanager.healthcareapis.models.PublicNetworkAccess;
 import com.azure.resourcemanager.healthcareapis.models.ServiceEventState;
 import com.azure.resourcemanager.healthcareapis.models.ServiceManagedIdentityIdentity;
+import com.azure.resourcemanager.healthcareapis.models.StorageConfiguration;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -101,6 +102,14 @@ public final class DicomServiceImpl implements DicomService, DicomService.Defini
         return this.innerModel().encryption();
     }
 
+    public StorageConfiguration storageConfiguration() {
+        return this.innerModel().storageConfiguration();
+    }
+
+    public Boolean enableDataPartitions() {
+        return this.innerModel().enableDataPartitions();
+    }
+
     public Region region() {
         return Region.fromName(this.regionName());
     }
@@ -174,9 +183,9 @@ public final class DicomServiceImpl implements DicomService, DicomService.Defini
         com.azure.resourcemanager.healthcareapis.HealthcareApisManager serviceManager) {
         this.innerObject = innerObject;
         this.serviceManager = serviceManager;
-        this.resourceGroupName = Utils.getValueFromIdByName(innerObject.id(), "resourceGroups");
-        this.dicomServiceName = Utils.getValueFromIdByName(innerObject.id(), "dicomservices");
-        this.workspaceName = Utils.getValueFromIdByName(innerObject.id(), "workspaces");
+        this.resourceGroupName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "resourceGroups");
+        this.dicomServiceName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "dicomservices");
+        this.workspaceName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "workspaces");
     }
 
     public DicomService refresh() {
@@ -244,6 +253,16 @@ public final class DicomServiceImpl implements DicomService, DicomService.Defini
 
     public DicomServiceImpl withEncryption(Encryption encryption) {
         this.innerModel().withEncryption(encryption);
+        return this;
+    }
+
+    public DicomServiceImpl withStorageConfiguration(StorageConfiguration storageConfiguration) {
+        this.innerModel().withStorageConfiguration(storageConfiguration);
+        return this;
+    }
+
+    public DicomServiceImpl withEnableDataPartitions(Boolean enableDataPartitions) {
+        this.innerModel().withEnableDataPartitions(enableDataPartitions);
         return this;
     }