+++ sdk/healthcareapis/azure-resourcemanager-healthcareapis/src/main/java/com/azure/resourcemanager/healthcareapis/implementation/FhirServiceImpl.java
@@ -200,9 +200,9 @@ public final class FhirServiceImpl implements FhirService, FhirService.Definitio
         com.azure.resourcemanager.healthcareapis.HealthcareApisManager serviceManager) {
         this.innerObject = innerObject;
         this.serviceManager = serviceManager;
-        this.resourceGroupName = Utils.getValueFromIdByName(innerObject.id(), "resourceGroups");
-        this.fhirServiceName = Utils.getValueFromIdByName(innerObject.id(), "fhirservices");
-        this.workspaceName = Utils.getValueFromIdByName(innerObject.id(), "workspaces");
+        this.resourceGroupName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "resourceGroups");
+        this.fhirServiceName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "fhirservices");
+        this.workspaceName = ResourceManagerUtils.getValueFromIdByName(innerObject.id(), "workspaces");
     }
 
     public FhirService refresh() {